Another tweak to player viewheight/lookdir on death

+++ src/p_user.c
@@ -279,7 +279,7 @@ static void P_DeathThink(void)
 
             if (deadlookdir == -1)
             {
-                float   viewheightrange = (viewplayer->viewheight - DEADVIEWHEIGHT) / FRACUNIT;
+                float   viewheightrange = (float)(viewplayer->viewheight - DEADVIEWHEIGHT) / FRACUNIT;
 
                 inc = MAX(1, ABS(DEADLOOKDIR - viewplayer->lookdir));
